requirement.txt
openai==2.8.1
playwright==1.55.0
agentql==1.15.0
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.2
pyairtable==3.2.0



streamlit
pandas
plotly
numpy
psycopg2-binary
sqlalchemy
redis
wordcloud
matplotlib
